Add files via upload
INTRODUCTION 

 Administrator and staff are a distributed application, developed to maintain the details of employees working in any organization. It maintains the information about the personal details of their employees, also the details about the payroll system which enable to generate the pays lip. The application is actually a suite of applications developed using Java. 
 
It is simple to understand and can be used by anyone who is not even familiar with simple employee system. It is user friendly and just asks the user to follow step by step operations by giving him few options. It is fast and can perform many operations of a company. This software package has been developed using the powerful coding tools of JAVA at Front End and Microsoft SQL Server at Back End. The software is very user friendly.  The package contains different modules like Employee details. This version of the software has multi-user approach. For further enhancement or development of the package, user’s feedback will be considered. 
 
Administrator and staff system are an application that enables users to create and store Employee Records. The application also provides facilities of a payroll system which enables user to generate Pay slips too. This application is helpful to department of the organization which maintains data of employees related to an organization. Java is a platform independent language. Its created applications can be used on a standalone machine as well as on distributed network. More over applications developed in java can be extended to Internet based applications.